+++ truffle/com.oracle.truffle.api/src/com/oracle/truffle/api/impl/Accessor.java
@@ -39,6 +39,7 @@ import com.oracle.truffle.api.frame.MaterializedFrame;
 import com.oracle.truffle.api.nodes.Node;
 import com.oracle.truffle.api.nodes.RootNode;
 import com.oracle.truffle.api.source.Source;
+import com.oracle.truffle.api.source.SourceSection;
 import java.util.Collection;
 
 /**
@@ -114,6 +115,10 @@ public abstract class Accessor {
         public abstract void postInitEnv(Env env);
 
         public abstract Object getVM(Env env);
+
+        public abstract Object findMetaObject(TruffleLanguage<?> language, Env env, Object value);
+
+        public abstract SourceSection findSourceLocation(TruffleLanguage<?> language, Env env, Object value);
     }
 
     public abstract static class InstrumentSupport {